+++ Application/DTOs/HomeWorkSubmitionDTO/Queries/GetAllHomeWorkForStudentViewModel.cs
@@ -21,8 +21,8 @@ namespace Application.DTOs
         public ApplicationUser CorrectionTeacher { get; set; }
         public string Status { get; set; }
         public DateTime? SubmitionDate { get; set; }
-        public int Points { get; set; }
+        public double Points { get; set; }
         public string Comment { get; set; }
-        public int BonusPoints { get; set; }
+        public double BonusPoints { get; set; }
     }
 }
